\name{regression.test.refdata}
\alias{regression.test.refdata}
\title{ regression test for refdata }
\description{
  This function checks a series of use cases.
}
\usage{
regression.test.refdata()
}
\details{
  raises a stop error if a problem is detected
}
\value{
  TRUE if successful
}
\author{ Jens Oehlschlägel }
\seealso{ \code{\link{refdata}}, \code{\link[utils]{example}} }

\examples{
  regression.test.refdata()
}
\keyword{ internal }
