\name{sleep.wrapper}
\alias{sleep.wrapper}
\alias{memsize.wrapper}
\alias{timing.wrapper}
\title{ wrapper to get some measures for all platforms }
\description{
  interrupts execution for specified no. of seconds
}
\usage{
sleep.wrapper(time)
}
\arguments{
  \item{time}{ no. of seconds }
}
\value{
  NULL
}
\author{ Jens Oehlschlägel }
\seealso{ \code{\link[base]{Sys.sleep}} }
\keyword{ internal }
